refactor(agent): split graph tier and memory orchestration out of component.ts
Address the thermos code-quality findings on the agent-memory feature by
extracting the graph tier and the memory-source resolution into their own
modules, shrinking component.ts from >1000 lines back under the size budget.
Pure structural change — no behavior change; the @lunora/agent/component
public subpath still re-exports agentComponent/agentExtension and its
functions map still exposes the internal-stamped agentGraphTraverse/
agentGraphUpsert.

- graph-component.ts (new): the entities/edges tables, pure BFS helpers, and
  the agentGraphUpsert/agentGraphTraverse registered functions.
- component-shared.ts (new): the small primitives both component modules need
  (AGENT_EXTENSION_KEY, asInternal, definedColumns, AgentRegisteredFunction).
- memory.ts (new): the single source of truth for which memory sources are
  auto-injected (isInjectedMemorySource) and the durable-step-name convention
  (memoryStepName), plus resolveInjectedSources/firstGraphSource. Byte-identical
  step names preserve in-flight workflow replay safety.
- agent-loop.ts: route retrieve/traverse/extract through the shared helpers.
- component.ts: re-export the graph pieces; keep only the thread/state/approval
  functions inline.
